Refactor branch list context menu callback to the list item
