+++ app/src/ui/branches/branch-renderer.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { Branch, BranchType } from '../../models/branch'
+import { Branch } from '../../models/branch'
 
 import { IBranchListItem } from './group-branches'
 import { BranchListItem } from './branch-list-item'
@@ -10,8 +10,6 @@ export function renderDefaultBranch(
   item: IBranchListItem,
   matches: IMatches,
   currentBranch: Branch | null,
-  onRenameBranch?: (branchName: string) => void,
-  onDeleteBranch?: (branchName: string) => void,
   onDropOntoBranch?: (branchName: string) => void,
   onDropOntoCurrentBranch?: () => void
 ): JSX.Element {
@@ -22,11 +20,8 @@ export function renderDefaultBranch(
     <BranchListItem
       name={branch.name}
       isCurrentBranch={branch.name === currentBranchName}
-      isLocal={branch.type === BranchType.Local}
       lastCommitDate={commit ? commit.author.date : null}
       matches={matches}
-      onRenameBranch={onRenameBranch}
-      onDeleteBranch={onDeleteBranch}
       onDropOntoBranch={onDropOntoBranch}
       onDropOntoCurrentBranch={onDropOntoCurrentBranch}
     />
