+++ app/src/ui/branches/branches-container.tsx
@@ -206,8 +206,6 @@ export class BranchesContainer extends React.Component<
       item,
       matches,
       this.props.currentBranch,
-      this.props.onRenameBranch,
-      this.props.onDeleteBranch,
       this.onDropOntoBranch,
       this.onDropOntoCurrentBranch
     )
@@ -239,6 +237,8 @@ export class BranchesContainer extends React.Component<
               DragType.Commit
             )}
             renderPreList={this.renderPreList}
+            onRenameBranch={this.props.onRenameBranch}
+            onDeleteBranch={this.props.onDeleteBranch}
           />
         )
